ref: move StockPage composable function to another directory
